+++ src/main/java/com/anvisa/model/persistence/mongodb/sequence/SequenceId.java
@@ -1,5 +1,7 @@
 package com.anvisa.model.persistence.mongodb.sequence;
 
+import java.math.BigInteger;
+
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
@@ -9,7 +11,7 @@ public class SequenceId {
 	@Id
 	private String id;
 
-	private long seq;
+	private BigInteger seq;
 
 	public String getId() {
 		return id;
@@ -19,11 +21,11 @@ public class SequenceId {
 		this.id = id;
 	}
 
-	public long getSeq() {
+	public BigInteger getSeq() {
 		return seq;
 	}
 
-	public void setSeq(long seq) {
+	public void setSeq(BigInteger seq) {
 		this.seq = seq;
 	}
